changed int by size_t
git-svn-id: file:///home/roberto/Git/svn@8865 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: e0426dd02d975f10f5d36e9456e55f010f5eec0c

+++ trunk/xmipp/applications/programs/angular_distribution_show/angular_distribution_show_main.cpp
@@ -89,7 +89,7 @@ public:
         // Get angles ==============================================================
         MetaData angles;
         angles.read(fnIn);
-        int AngleNo = angles.size();
+        size_t AngleNo = angles.size();
         if (AngleNo == 0 || !angles.containsLabel(MDL_ANGLEROT))
             REPORT_ERROR(ERR_MD_BADLABEL, "Input file doesn't contain angular information");
